+++ assets/frontend/router/index.ts
@@ -23,5 +23,5 @@ export const routes: RouteRecordRaw[] = [
     { path: "/whats-new", component: () => import("../pages/NewsPage.vue") },
     { path: "/profile", component: () => import("../pages/UserProfilePage.vue") },
     { path: "/api-tokens", component: () => import("../pages/ApiTokensPage.vue") },
-    { path: "/:pathMatch(.*)*", component: NotFoundPage },
+    { path: "/:pathMatch(.*)*", name: "not-found", component: NotFoundPage },
 ];